review fixes: moved supplementary functions from ui-text to foundation, removed changes in public API (ui-text), refactor logic in TextFieldDelegate.skiko.kt, added documentation to some supplementary functions
