+++ compose/foundation/foundation/src/skikoMain/kotlin/androidx/compose/foundation/text/StringHelpers.skiko.kt
@@ -29,3 +29,156 @@ internal actual fun String.findFollowingBreak(index: Int): Int {
     it.setText(this)
     return it.following(index)
 }
+
+// TODO Remove once it's available in common stdlib https://youtrack.jetbrains.com/issue/KT-23251
+internal typealias CodePoint = Int
+
+/**
+ * Converts a surrogate pair to a unicode code point.
+ */
+private fun Char.Companion.toCodePoint(high: Char, low: Char): CodePoint =
+    (((high - MIN_HIGH_SURROGATE) shl 10) or (low - MIN_LOW_SURROGATE)) + 0x10000
+
+/**
+ * The minimum value of a supplementary code point, `\u0x10000`.
+ */
+private const val MIN_SUPPLEMENTARY_CODE_POINT: Int = 0x10000
+
+internal fun CodePoint.charCount(): Int = if (this >= MIN_SUPPLEMENTARY_CODE_POINT) 2 else 1
+
+internal val String.codePoints
+    get() = sequence {
+        var index = 0
+        while (index < length) {
+            val codePoint = codePointAt(index)
+            yield(codePoint)
+            index += codePoint.charCount()
+        }
+    }
+
+/**
+ * Returns the character (Unicode code point) at the specified index.
+ */
+internal fun String.codePointAt(index: Int): CodePoint {
+    val high = this[index]
+    if (high.isHighSurrogate() && index + 1 < this.length) {
+        val low = this[index + 1]
+        if (low.isLowSurrogate()) {
+            return Char.toCodePoint(high, low)
+        }
+    }
+    return high.code
+}
+
+
+/**
+ * Finds the offset of the next non-whitespace symbols subsequence (word) in the given text
+ * starting from the specified caret offset.
+ *
+ * @param offset The offset where to start looking for the next word.
+ * @param currentText The current text in which to search for the next word.
+ * @return The offset of the next non-whitespace symbols subsequence (word), or the end of the string
+ *         if no such word is found.
+ */
+internal fun findNextNonWhitespaceSymbolsSubsequenceStartOffset(
+    offset: Int,
+    currentText: String
+): Int {
+    /* Assume that next non whitespaces symbols subsequence (word) is when current char is whitespace and next character is not.
+     * Emoji (compound incl.) should be treated as a new word.
+     */
+    val charIterator =
+        BreakIterator.makeCharacterInstance() // wordInstance doesn't consider symbols sequence as word
+    charIterator.setText(currentText)
+
+    var currentOffset: Int
+    var nextOffset = charIterator.next()
+    while (nextOffset < offset) {
+        nextOffset = charIterator.next()
+    }
+    currentOffset = nextOffset
+
+    while (nextOffset != BreakIterator.DONE) {
+        nextOffset = charIterator.next()
+        if (currentText.codePointAt(currentOffset).isWhitespace() && !currentText.codePointAt(
+                nextOffset
+            ).isWhitespace()
+        ) {
+            return currentOffset
+        } else {
+            currentOffset = nextOffset
+        }
+    }
+    return currentOffset
+}
+
+/**
+ * Determines whether the character at the specified offset in the string is a whitespace Unicode character.
+ *
+ * @param offset The index of the character to check.
+ * @return `true` if the character at the specified offset is a whitespace character, `false` otherwise.
+ */
+internal fun String.isWhitespace(offset: Int): Boolean {
+    return this.codePointAt(offset).isWhitespace()
+}
+
+/**
+ * Checks if the character at the specified offset in the string is a punctuation Unicode character.
+ *
+ * @param offset The offset of the character to check.
+ * @return true if the character at the specified offset is a punctuation character, false otherwise.
+ */
+internal fun String.isPunctuation(offset: Int): Boolean {
+    return this.codePointAt(offset).isPunctuation()
+}
+
+internal fun String.halfSymbolsOffset(): Int {
+    val symbolsCount = this.codePoints.count()
+    val charIterator = BreakIterator.makeCharacterInstance()
+    charIterator.setText(this)
+    var currentOffset = 0
+    for (i in 0..symbolsCount / 2) {
+        currentOffset = charIterator.next()
+    }
+    return currentOffset
+}
+
+/**
+ * Checks if the given Unicode code point is a whitespace character.
+ *
+ * @return `true` if the code point is a whitespace character, `false` otherwise.
+ */
+private fun CodePoint.isWhitespace(): Boolean {
+    // TODO: Extend this behavior when (if) Unicode will have compound whitespace characters.
+    if (this.charCount() != 1) {
+        return false
+    }
+    return this.toChar().isWhitespace()
+}
+
+/**
+ * Checks if the given Unicode code point is a punctuation character.
+ *
+ * @return 'true' if the CodePoint is a punctuation character, 'false' otherwise.
+ */
+private fun CodePoint.isPunctuation(): Boolean {
+    // TODO: Extend this behavior when (if) Unicode will have compound punctuation characters.
+    if (this.charCount() != 1) {
+        return false
+    }
+    val punctuationSet = setOf(
+        CharCategory.DASH_PUNCTUATION,
+        CharCategory.START_PUNCTUATION,
+        CharCategory.END_PUNCTUATION,
+        CharCategory.CONNECTOR_PUNCTUATION,
+        CharCategory.OTHER_PUNCTUATION,
+        CharCategory.INITIAL_QUOTE_PUNCTUATION,
+        CharCategory.FINAL_QUOTE_PUNCTUATION
+    )
+    punctuationSet.forEach {
+        if (it.contains(this.toChar())) {
+            return true
+        }
+    }
+    return false
+}
